Convert Ratpack interceptor to initializer
This solves an issue where multiple execution segments would try to clean up
the same profiler, despite being the same execution.

We'll also now close the profiler on execution complete if that happens before
the response is sent, when there's a response. Previously we'd wait until the
response was about to be sent, even if the execution completed first.